Refactor Notepad application and update dependencies
- Updated Cargo.lock and Cargo.toml to replace `uisys_client` with `dynlink` and `libanyui_client`.
- Refactored main.rs to utilize `libanyui` for UI components, including toolbar and status bar.
- Removed the custom Editor struct and replaced it with `anyui::TextEditor`.
- Implemented event handling for toolbar buttons and keyboard shortcuts.
- Enhanced context menu handling in libanyui to support hover effects and right-click actions.
- Improved DataGrid control to correctly draw column separators based on content height.
- Updated event loop to use time-based double-click detection instead of frame-based.
- Added functionality to the Finder application for right-click context menus and keyboard navigation in icon view.

+++ libs/libanyui/src/lib.rs
@@ -128,8 +128,8 @@ pub(crate) struct AnyuiState {
     pub hovered: Option<ControlId>,
     /// Control ID of the last click (for double-click detection).
     pub last_click_id: Option<ControlId>,
-    /// Tick count of the last click (for double-click detection).
-    pub last_click_tick: u64,
+    /// Timestamp (ms) of the last click (for double-click detection).
+    pub last_click_tick: u32,
     /// Which mouse button was pressed (for right-click detection).
     pub pressed_button: u32,
